requirements.txt
git-python==1.0.3
gitdb==4.0.11
GitPython==3.1.42
smmap==5.0.1
